requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
chromadb==0.4.18
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
python-dotenv==1.0.0
